requirements-production.txt
# Core Flask dependencies
Flask==2.3.3
Flask-WTF==1.2.1
Flask-Cors==6.0.1
WTForms==3.0.1
python-dotenv==1.0.0
redis==5.0.1
Flask-Login==0.6.3
Authlib==1.3.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Mail==0.10.0
PyJWT==2.8.0
email-validator>=2.0.0
stripe==7.8.1

# YouTube API and HTTP requests
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
requests==2.31.0

# Data processing (for analytics/reporting)
pandas==2.1.4
numpy==1.26.2

# Utilities
urllib3==1.26.18
python-dateutil==2.8.2
httpx==0.27.0

# Production server
gunicorn==21.2.0

# PostgreSQL driver (binary)
psycopg2-binary==2.9.9
psycopg[binary]==3.2.1
